+++ sarpy/utils/convert_to_sicd.py
@@ -8,7 +8,7 @@ For a basic help on the command-line, check
 """
 
 __classification__ = "UNCLASSIFIED"
-__author__ = "Thomas McCullough"
+__author__ = ("Thomas McCullough", "Valkyrie Systems Corporation")
 
 import argparse
 import logging
@@ -17,7 +17,8 @@ import sarpy
 from sarpy.io.complex.converter import conversion_utility
 
 
-def convert(input_file, output_dir, preserve_nitf_information=False):
+def convert(input_file, output_dir, preserve_nitf_information=False,
+            dem_filename_pattern=None, dem_type=None, geoid_file=None):
     """
 
     Parameters
@@ -29,17 +30,53 @@ def convert(input_file, output_dir, preserve_nitf_information=False):
     preserve_nitf_information : bool
         Try to preserve NITF information? This only applies in the case that the
         file being read is actually a NITF file.
+    dem_filename_pattern : str | None
+        Optional string specifying a Digital Elevation Model (DEM) filename pattern.
+        This is a format string that specifies a glob pattern that will
+        uniquely specify a DEM file from the Lat/Lon of the SW corner of
+        the DEM tile.  See the convert_to_sicd help text for more details.
+    dem_type : str | None
+        Optional DEM type ('GeoTIFF', etc.).
+        This parameter is required when dem_filename_pattern is specified.
+    geoid_file : str | None
+        Optional Geoid file which might be needed when dem_filename_pattern is specified.
     """
 
-    conversion_utility(input_file, output_dir, preserve_nitf_information=preserve_nitf_information)
+    conversion_utility(input_file, output_dir, preserve_nitf_information=preserve_nitf_information,
+                       dem_filename_pattern=dem_filename_pattern, dem_type=dem_type, geoid_file=geoid_file)
 
 
 if __name__ == '__main__':
+    epilog = ('Note:\n'
+              'The DEM files must have the SW corner Lat/Lon encoded in their filenames.\n'
+              'The --dem-path-pattern argument contains a format string that when populated will\n'
+              'create as glob pattern that will specify the desired DEM file.  The following\n'
+              'arguments are provided to the format string.\n'
+              '    lat = int(numpy.floor(lat))\n'
+              '    lon = int(numpy.floor(lon))\n'
+              '    abslat = int(abs(numpy.floor(lat)))\n'
+              '    abslon = int(abs(numpy.floor(lon)))\n'
+              '    ns = "s" if lat < 0 else "n"\n'
+              '    NS = "S" if lat < 0 else "N"\n'
+              '    ew = "w" if lon < 0 else "e"\n'
+              '    EW = "W" if lon < 0 else "E"\n'
+              '\n'
+              'For example (with Linux file separators) the following specifies a GeoTIFF DEM pattern:\n'
+              '    /dem_root/tdt_{ns}{abslat:02}{ew}{abslon:03}_*/DEM/TDT_{NS}{abslat:02}{EW}{abslon:03}_*_DEM.tif\n'
+              '\n'
+              'In theory, one could use a simple format string using wildcard characters like this:\n'
+              '    /dem_root/**/*{NS}{abslat:02}{EW}{abslon:03}*DEM.tif\n'
+              '\n'
+              'However, this would be unwise since glob might have to scan through many files and directories\n'
+              'to find the desired file.  This could be quite time consuming if there are many files in dem_root.\n'
+              )
+
     parser = argparse.ArgumentParser(description="Convert to SICD format.",
-                                     formatter_class=argparse.RawTextHelpFormatter)
+                                     formatter_class=argparse.RawTextHelpFormatter,
+                                     epilog=epilog)
     parser.add_argument(
         'input_file', metavar='input_file',
-        help='Path input data file, or directory for radarsat, RCM, or sentinel.\n'
+        help='Path input data file, or directory for radarsat, RCM, sentinel or other systems.\n'
              '* For radarsat or RCM, this can be the product.xml file, or parent directory\n'
              '  of product.xml or metadata/product.xml.\n'
              '* For sentinel, this can be the manifest.safe file, or parent directory of\n'
@@ -55,6 +92,21 @@ if __name__ == '__main__':
         help='Try to preserve any NITF information?\n'
              'This only applies in the event that the file being read is a NITF')
     parser.add_argument(
+        '-d', '--dem-filename-pattern',
+        help='Optional string specifying a Digital Elevation Model (DEM) filename pattern.\n'
+             'This is a format string that specifies a glob pattern that will\n'
+             'uniquely specify a DEM file from the Lat/Lon of the SW corner of\n'
+             'the DEM tile.  See the note below for more details.\n')
+    parser.add_argument(
+        '-t', '--dem-type',
+        help=('Optional DEM type ("GeoTIFF", etc.).\n'
+              'This parameter is required when dem-path-pattern is specified.\n'))
+    parser.add_argument(
+        '-g', '--geoid-file',
+        help='Optional path to a geoid definition file.\n'
+             'A geoid definition file is required when dem-path-pattern is specified\n'
+             'and the DEM height values are relative to a geoid.\n')
+    parser.add_argument(
         '-v', '--verbose', action='store_true', help='Verbose (level="INFO") logging?')
 
     args = parser.parse_args()
@@ -63,4 +115,5 @@ if __name__ == '__main__':
     logger = logging.getLogger('sarpy')
     logger.setLevel(level)
 
-    convert(args.input_file, args.output_directory, preserve_nitf_information=args.preserve)
+    convert(args.input_file, args.output_directory, preserve_nitf_information=args.preserve,
+            dem_filename_pattern=args.dem_filename_pattern, dem_type=args.dem_type, geoid_file=args.geoid_file)